--- Tools/bin/exportBlah
@@ -1,13 +1,0 @@
-# !/usr/bin/env python
-
-import sys
-#sys.path.insert(1, "/home/ross/EuPathGalaxy/Tools/lib/python/")
-sys.path.insert(0, "/opt/galaxy/tools/eupath/Tools/lib/python")
-from eupath import GeneListEupathExporter
-from eupath import EupathExporterNew
-
-def main():
-    EupathExporterNew.execute(GeneListEupathExporter.GeneListExporter())
-
-if __name__ == "__main__":
-    sys.exit(main())
