+++ Tools/lib/python/eupath/EupathExporter.py
@@ -6,362 +6,96 @@ import time
 import os
 import shutil
 import sys
+import time
 import requests
 import tempfile
 import contextlib
 import re
-from requests.auth import HTTPBasicAuth
-from subprocess import Popen, PIPE
+import optparse
 
+def print_debug(msg):
+    if os.getenv('DEBUG'):
+        print(msg,  file=sys.stderr)
+    
+def execute(exporter):
+    (options, args) = optparse.OptionParser().parse_args()
+    stdArgsBundle = StandardArgsBundle(args)
+    typeSpecificArgsList = stdArgsBundle.getTypeSpecificArgsList(args)
+    exporter.initialize(stdArgsBundle, typeSpecificArgsList);
 
-class Export:
-    """
-    This is a generic VEuPathDB export tool for Galaxy.  It is abstract and so must be subclassed by more
-    specialized export tools that implement those abstract classes.
-    """
+    try:
+        print_debug("Attempting export.")
+        exporter.export()
+    except SystemException as ve:
+        print(str(ve), file=sys.stderr)
+        sys.exit(1)
 
-    # Names for the 2 json files and the folder containing the dataset to be included in the tarball
-    DATASET_JSON = "dataset.json"
-    META_JSON = "meta.json"
-    DATAFILES = "datafiles"
+class StandardArgsBundle:
+    ARGS_LEN = 6
 
-    def __init__(self, dataset_type, version, validation_script, args):
-        """
-        Initializes the export class with the parameters needed to accomplish the export of user
-        datasets on Galaxy to VEuPathDB projects.
-        :param dataset_type: The VEuPathDB type of this dataset
-        :param version: The version of the VEuPathDB type of this dataset
-        :param validation_script: A script that handles the validation of this dataset
-        :param args: An array of the input parameters
-        """
-        self._type = dataset_type
-        self._version = version
-        self._validation_script = validation_script
+    def __init__(self, args):
+        if len(args) < StandardArgsBundle.ARGS_LEN:
+            raise SystemException("The export tool was passed an insufficient numbers of arguments.")
+        self.dataset_name = args[0]
+        self.summary = args[1]
+        self.description = args[2]
+        self.user_id = getWdkUserId(args[3])
+        self.tool_directory = args[4] # Used to find the configuration file containing IRODS url and credentials
+        self.output = args[5] # output file
 
-        # Extract and transform the parameters as needed into member variables
-        self.parse_params(args)
+    def getTypeSpecificArgsList(self, args):
+        return args[StandardArgsBundle.ARGS_LEN : ]
 
-        # This msec timestamp is used to denote both the created and modified times.
-        self._timestamp = int(time.time() * 1000)
+# An abstract class to export to VEuPathDB.  Subclasses implement details for a given dataet type
+class Exporter:
+    POLLING_FACTOR = 1.5  # multiplier for progressive polling of status endpoint
+    POLLING_INTERVAL_MAX = 60
+    POLLING_TIMEOUT = 10 * POLLING_INTERVAL_MAX 
+    SOURCE_GALAXY = "galaxy" # indicate to the service that Galaxy is the point of origin for this user dataset.
 
-        # This is the name of the file to be exported sans extension.  It will be used to designate a unique temporary
-        # directory and to export both the tarball and the flag that triggers IRODS to process the tarball. By
-        # convention, the dataset tarball is of the form dataset_uNNNNNN_tNNNNNNN.tgz where the NNNNNN following the _u
-        # is the WDK user id and _t is the msec timestamp
-        self._export_file_root = 'dataset_u' + str(self._user_id) + '_t' + str(self._timestamp) + '_p' + str(os.getpid())
-        print("Export file root is " + self._export_file_root, file=sys.stdout)
+    def initialize(self, stdArgsBundle, dataset_type, dataset_version):
+        self._stdArgsBundle = stdArgsBundle
+        self._dataset_type = dataset_type
+        self._dataset_version = dataset_version
 
-        # Set up the configuration data
-        (self._url, self._user, self._pwd, self._lz_coll, self._flag_coll) = self.collect_rest_data()
+        # create a unique name for our tmp working dir and the tarball, of the form: 
+        #   dataset_uNNNNNN_tTTTTTTT 
+        # where NNNNNN is the VEupath user id and TTTTTT is the msec timestamp
+        timestamp = int(time.time() * 1000)
+        self._export_file_root = 'dataset_u' + str(self._stdArgsBundle.user_id) + '_t' + str(timestamp) + '_p' + str(os.getpid())
+        print_debug("Export file root is " + self._export_file_root)
 
-    def parse_params(self, args):
-        """
-        Salts away all generic parameters (i.e., the first 5 params) and do some initial validation.  The subclasses
-        will handle the other parameters.
-        :param args:
-        :return:
-        """
-        if len(args) < 6:
-            raise ValidationException("The tool was passed an insufficient numbers of arguments.")
-        self._dataset_name = args[0]
-        self._summary = args[1]
-        self._description = args[2]
+        # read in config info
+        (self._url, self._user_id, self._pwd, self._service_url, self._super_user_token) = self.read_config()
 
-        # WDK user id is derived from the user email
-        user_email = args[3].strip()
-        if not re.match(r'.+\.\d+@veupathdb.org$', user_email, flags=0):
-            raise ValidationException(
-                "The user email " + str(user_email) + " is not valid for the use of this tool.")
-        galaxy_user = user_email.split("@")[0]
-        self._user_id = galaxy_user[galaxy_user.rfind(".") + 1:]
-
-        # Used to find the configuration file containing IRODS url and credentials
-        self._tool_directory = args[4]
-
-        # Output file
-        self._output = args[5]
-
-
-    def collect_rest_data(self):
+    def read_config(self):
         """
         Obtains the url and credentials and relevant collections needed to run the iRODS rest service.
         At some point, this information should be fished out of a configuration file.
-        :return:  A tuple containing the url, user, and password, landing zone and flags collection,
-         in that order
         """
-        config_path = self._tool_directory + "/../../config/config.json"
+        config_path = self._stdArgsBundle.tool_directory + "/../../config/config.json"
         
-        # The tool directory path seems glitchy on Globus Dev Galaxy instance after service restarts.
-        # Uncomment to check.
-        #print >> sys.stdout, "self._tool_directory is " + self._tool_directory
         with open(config_path, "r") as config_file:
             config_json = json.load(config_file)
-            return (config_json["url"], config_json["user"], config_json["password"], "lz", "flags")
-
-    def validate_datasets(self):
-        """
-        Runs the validation script provided to the class upon initialization using the user's
-        dataset files as standard input.
-        :return:
-        """
-
-        if self._validation_script == None:
-            return
-        
-        dataset_files = self.identify_dataset_files()
-
-        validation_process = Popen(['python', self._tool_directory + "/../../bin/" + self._validation_script],
-                                   stdin=PIPE, stdout=PIPE, stderr=PIPE)
-        # output is a tuple containing (stdout, stderr)
-        output = validation_process.communicate(json.dumps(dataset_files).encode("utf-8"))
-        if validation_process.returncode == 1:
-            raise ValidationException(output[1])
-
-    def identify_dependencies(self):
-        """
-        An abstract method to be addressed by a specialized export tool that furnishes a dependency json list.
-        :return: The dependency json list to be returned should look as follows:
-        [dependency1, dependency2, ... ]
-        where each dependency is written as a json object as follows:
-        {
-          "resourceIdentifier": <value>,
-          "resourceVersion": <value>,
-          "resourceDisplayName": <value
-        }
-        Where no dependencies exist, an empty list is returned
-        """
-        raise NotImplementedError(
-            "The method 'identify_dependencies(self)' needs to be implemented in the specialized export module.")
-
-    def identify_projects(self):
-        """
-        An abstract method to be addressed by a specialized export tool that furnishes a VEuPathDB project list.
-        :return: The project list to be returned should look as follows:
-        [project1, project2, ... ]
-        At least one valid VEuPathDB project must be listed
-        """
-        raise NotImplementedError(
-            "The method 'identify_project(self)' needs to be implemented in the specialized export module.")
-
-    def identify_supported_projects(self):
-        """
-        Override this method to provide a non-default list of projects.
-
-        Default is None, interpreted as all projects are ok, ie, no constraints.
-        """
-        return None;
-
-
-    def identify_dataset_files(self):
-        """
-        An abstract method to be addressed by a specialized export tool that furnishes a json list
-        containing the dataset data files and the VEuPathDB file names they must have in the tarball.
-        :return: The dataset file list to be returned should look as follows:
-        [dataset file1, dataset file2, ... ]
-        where each dataset file is written as a json object as follows:
-        {
-          "name":<filename that VEuPathDB expects>,
-          "path":<Galaxy path to the dataset file>
-        At least one valid VEuPathDB dataset file must be listed
-        """
-        raise NotImplementedError(
-            "The method 'identify_dataset_file(self)' needs to be implemented in the specialized export module.")
-
-    def create_dataset_json_file(self, temp_path):
-        """ Create and populate the dataset.json file that must be included in the tarball."""
-
-        # Get the total size of the dataset files (needed for the json file)
-        size = sum(os.stat(dataset_file['path']).st_size for dataset_file in self.identify_dataset_files())
-
-        if self.identify_supported_projects() != None:
-            for (project) in self.identify_projects():
-                if project not in self.identify_supported_projects():
-                    raise ValidationException("Sorry, you cannot export this kind of data to " + project)
-
-        dataset_path = temp_path + "/" + self.DATASET_JSON
-        with open(dataset_path, "w+") as json_file:
-            json.dump({
-              "type": {"name": self._type, "version": self._version},
-              "dependencies": self.identify_dependencies(),
-              "projects": self.identify_projects(),
-              "dataFiles": self.create_data_file_metadata(),
-              "owner": self._user_id,
-              "size": size,
-              "created": self._timestamp
-            }, json_file, indent=4)
-
-    def create_metadata_json_file(self, temp_path):
-        """" Create and populate the meta.json file that must be included in the tarball."""
-        meta_path = temp_path + "/" + self.META_JSON
-        with open(meta_path, "w+") as json_file:
-            json.dump({"name": self._dataset_name,
-                       "summary": self._summary,
-                       "description": self._description
-                       }, json_file, indent=4)
-
-    def create_data_file_metadata(self):
-        """
-        Create a json object holding metadata for an array of dataset files.
-        :return: json object to be inserted into dataset.json
-        """
-        dataset_files_metadata = []
-        for dataset_file in self.identify_dataset_files():
-            dataset_file_metadata = {}
-            dataset_file_metadata["name"] = self.clean_file_name(dataset_file['name'])
-            dataset_file_metadata["file"] = os.path.basename(dataset_file['path'])
-            dataset_file_metadata["size"] = os.stat(dataset_file['path']).st_size
-            dataset_files_metadata.append(dataset_file_metadata)
-        return dataset_files_metadata
-
-    
-    # replace undesired characters with underscore
-    def clean_file_name(self, file_name):
-        s = str(file_name).strip().replace(' ', '_')
-        return re.sub(r'(?u)[^-\w.]', '_', s)
-        
-    def package_data_files(self, temp_path):
-        """
-        Copies the user's dataset files to the datafiles folder of the temporary dir and changes each
-        dataset filename conferred by Galaxy to a filename expected by VEuPathDB
-        """
-        os.mkdir(temp_path + "/" + self.DATAFILES)
-        for dataset_file in self.identify_dataset_files():
-            shutil.copy(dataset_file['path'], temp_path + "/" + self.DATAFILES + "/" + self.clean_file_name(dataset_file['name']))
-
-    def create_tarball(self):
-        """
-        Package the tarball - contains meta.json, dataset.json and a datafiles folder containing the
-        user's dataset files
-        """
-        with tarfile.open(self._export_file_root + ".tgz", "w:gz") as tarball:
-            for item in [self.META_JSON, self.DATASET_JSON, self.DATAFILES]:
-                tarball.add(item)
-
-    def process_request(self, collection, source_file):
-        """
-        This method wraps the iRODS rest request into a try/catch to insure that bad responses are
-        reflected back to the user.
-        :param collection: the name of the workspaces collection to which the file is to be uploaded
-        :param source_file: the name of the file to be uploaded to iRODS
-        """
-        rest_response = self.send_request(collection, source_file)
-        try:
-            rest_response.raise_for_status()
-        except requests.exceptions.HTTPError as e:
-            print("Error: " + str(e), file=sys.stderr)
-            sys.exit(1)
-
-    def send_request(self, collection, source_file):
-        """
-        This request is intended as a multi-part form post containing one file to be uploaded.  iRODS Rest
-        does an iput followed by an iget, apparently.  So the response can be used to insure proper
-        delivery.
-        :param collection: the name of the workspaces collection to which the file is to be uploaded
-        :param source_file: the name of the file to be uploaded to iRODS
-        :return: the http response from an iget of the uploaded file
-        """
-        request = self._url + collection + "/" + source_file
-        headers = {"Accept": "application/json"}
-        upload_file = {"uploadFile": open(source_file, "rb")}
-        auth = HTTPBasicAuth(self._user, self._pwd)
-        try:
-            response = requests.post(request, auth=auth, headers=headers, files=upload_file)
-            response.raise_for_status()
-        except Exception as e:
-            print("Error: The dataset export could not be completed at this time.  The VEuPathDB" \
-                                 " workspace may be unavailable presently. " + str(e), file=sys.stderr)
-            sys.exit(2)
-        return response
-
-    def get_flag(self, collection, source_file):
-        """
-        This method picks up any flag (success or failure) from the flags collection in iRODs related to the dataset
-        exported to determine whether the export was successful.  If not, the nature of the failure is reported to the
-        user.  The failure report will normally be very general unless the problem is one that can possibly be remedied
-        by the user (e.g., going over quota).
-        :param collection: The iRODS collection holding the status flags
-        :param source_file: The dataset tarball name sans extension
-        """
-        time.sleep(5)  # arbitrary wait period before one time check for a flag.
-        auth = HTTPBasicAuth(self._user, self._pwd)
-        # Look for the presence of a success flag first and if none found, check for a failure flag.  If neither
-        # found, assume that to be a failure also.
-        try:
-            request = self._url + collection + "/" + "success_" + source_file
-            success = requests.get(request, auth=auth, timeout=5)
-            if success.status_code == 404:
-                request = self._url + collection + "/" + "failure_" + source_file
-                failure = requests.get(request, auth=auth, timeout=5)
-                if failure.status_code != 404:
-                    raise TransferException(failure.content)
-                else:
-                    failure.raise_for_status()
-            else:
-                self.output_success()
-                print("Your dataset has been successfully exported to VEuPathDB.", file=sys.stdout)
-                print("Please visit an appropriate VEuPathDB site to view your dataset.", file=sys.stdout)
-        except (requests.exceptions.ConnectionError, TransferException) as e:
-            print("Error: " + str(e), file=sys.stderr)
-            sys.exit(1)
-        
-    def connection_diagnostic(self):
-        """
-        Used to insure that the calling ip is the one expected (i.e., the one for which the
-        firewall is opened).  In Globus Dev Galaxy instance calling the tool outside of Galaxy
-        versus inside Galaxy resulted in different calling ip addresses.
-        """
-        request = "http://ifconfig.co"
-        headers = {"Accept": "application/json"}
-        try:
-            response = requests.get(request, headers=headers)
-            response.raise_for_status()
-            print("Diagnostic Result: " + response.content, file=sys.stdout)
-        except Exception as e:
-            print("Diagnostic Error: " + str(e), file=sys.stderr)        
-
+            return (config_json["url"], config_json["user"], config_json["password"], config_json["service-url"], config_json["super-user-token"])
+                    
     def export(self):
-        """
-        Does the work of exporting to VEuPathDB, a tarball consisting of the user's dataset files along
-        with dataset and metadata json files.
-        """
 
-        # Apply the validation first.  If it fails, exit with a data error.
-        self.validate_datasets()
-
-        # We need to save the current working directory so we can get back to it when we are
-        # finished working in our temporary directory.
+        # Save the current working directory so we can get back to it
         orig_path = os.getcwd()
 
-        # We need to create a temporary directory in which to assemble the tarball.
         with self.temporary_directory(self._export_file_root) as temp_path:
-
-            # Need to temporarily work inside the temporary directory to properly construct and
-            # send the tarball
             os.chdir(temp_path)
-
-            self.package_data_files(temp_path)
-            self.create_metadata_json_file(temp_path)
-            self.create_dataset_json_file(temp_path)
-            self.create_tarball()
-            
-            # Uncomment to check the calling ip address for this tool.
-            # self.connection_diagnostic()
-
-            # Call the iRODS rest service to drop the tarball into the iRODS workspace landing zone
-            self.process_request(self._lz_coll, self._export_file_root + ".tgz")
-
-            # Create a empty (flag) file corresponding to the tarball
-            open(self._export_file_root + ".txt", "w").close()
-
-            # Call the iRODS rest service to drop a flag into the IRODS workspace flags collection.  This flag
-            # triggers the iRODS PEP that unpacks the tarball and posts the event to Jenkins
-            self.process_request(self._flag_coll, self._export_file_root + ".txt")
-
-            # Look for a success/fail indication from IRODS.
-            self.get_flag(self._flag_coll, self._export_file_root)
-
-            # We exit the temporary directory prior to removing it, back to the original working directory.
-            os.chdir(orig_path)
+            print_debug("temp path: " + temp_path)
+            self.prepare_data_files(temp_path)
+            tarball_name = self.create_tarball(temp_path)
+            json_body = self.create_body_for_post()
+            print_debug(json_body)
+            user_dataset_id = self.post_metadata_json(json_body)
+            print_debug("UD ID: " + user_dataset_id)
+            self.post_datafile(user_dataset_id, tarball_name)
+            self.poll_for_upload_complete(user_dataset_id)   # teriminates if system or validation error
+            os.chdir(orig_path) # exit temp dir, prior to removing it
 
     @contextlib.contextmanager
     def temporary_directory(self, dir_name):
@@ -379,14 +113,122 @@ class Export:
             # Globus Dev Galaxy instance and it will throw an Exception if the boolean, 'True', is not in place.
             shutil.rmtree(temp_path, True)
 
+    def prepare_data_files(self, temp_path):
+        """
+        Copies the user's dataset files to the datafiles folder of the temporary dir and changes each
+        dataset filename conferred by Galaxy to a filename expected by VEuPathDB
+        """
+        for dataset_file in self.identify_dataset_files():
+            clean_name = temp_path + "/" + self.clean_file_name(dataset_file['name'])
+            print_debug("Creating dataset file: " + clean_name)
+            shutil.copy(dataset_file['path'], clean_name)
+
+    # replace undesired characters with underscore
+    def clean_file_name(self, file_name):
+        s = str(file_name).strip().replace(' ', '_')
+        return re.sub(r'(?u)[^-\w.]', '_', s)
+        
+    def create_tarball(self, temp_path):
+        """
+        Package the tarball - contains meta.json, dataset.json and a datafiles folder containing the
+        user's dataset files
+        """
+        tarball_name = self._export_file_root + ".tgz"
+        with tarfile.open(tarball_name, "w:gz") as tarball:
+            for filename in os.listdir(temp_path):
+                print_debug("Adding file to tarball: " + filename)
+                tarball.add(filename)
+        return tarball_name       
+
+    def create_body_for_post(self):
+        return {
+            "datasetName": self._stdArgsBundle.dataset_name,
+            "summary": self._stdArgsBundle.summary,
+            "description": self._stdArgsBundle.description,
+            "datasetType": self._datatype,
+            "projects": self.identify_projects(),
+            "origin": self.SOURCE_GALAXY
+        }
+
+    def post_metadata_json(self, json_blob):
+        headers = {"Accept": "application/json", "Auth-Key": self._super_user_token}
+        try:
+            response = requests.post(self._service_url, json = json_blob, headers=headers)
+            response.raise_for_status()
+            print_debug(response.json())
+            return response.json()['jobId']
+        except Exception as e:
+            self.printHttpErr("POST of metadata failed. " + str(e), response.status_code)            
+            print("Reason: " + response.text, file=sys.stderr)
+            sys.exit(1)
+
+    def post_datafile(self, user_dataset_id, tarball_name):
+        print_debug("POSTING data.  Tarball name: " + tarball_name)
+        headers = {"Accept": "application/json", "Auth-Key": self._super_user_token, "Originating-User_Id": self._user_id}
+        response = None
+        try:
+            form_fields = {"file": open(tarball_name, "rb"), "uploadMethod":"file"}
+            response = requests.post(self._service_url + "/" + user_dataset_id, headers=headers, files=form_fields)
+            response.raise_for_status()
+        except Exception as e:
+            self.printHttpErr("POST of metadata failed. " + str(e), response.status_code)            
+            if response != None:
+                print("Reason: " + response.text, file=sys.stderr)
+            sys.exit(1)
+
+    def poll_for_upload_complete(self, user_dataset_id):
+        start_time = time.time()
+        poll_interval = 1
+        while (self.check_upload_in_progress(user_dataset_id)):
+            time.sleep(poll_interval)  # sleep for specified seconds
+            if poll_interval < self.POLLING_INTERVAL_MAX:
+                poll_interval *= self.POLLING_FACTOR
+            if (time.time() - start_time > self.POLLING_TIMEOUT):
+                raise SystemException("Timed out polling for upload completion status")
+
+    # return True if still in progress; False if success.  Fail and terminate if system or validation error
+    def check_upload_in_progress(self, user_dataset_id):
+        headers = {"Accept": "application/json", "Auth-Key": self._super_user_token, "Originating-User_Id": self._user_id}
+        print_debug("Polling for status")
+        try:
+            response = requests.get(self._service_url + "/" + user_dataset_id, headers=headers)
+            response.raise_for_status()
+            json_blob = response.json()
+            if json_blob["status"] == "success":
+                return False
+            if json_blob["status"] == "errored":
+                self.handle_job_error_status(json_blob)
+            if json_blob["status"] == "rejected":
+                self.handle_job_rejected_status(json_blob)
+            return True
+        except Exception as e:
+            self.printHttpErr("GET of upload status failed. " + str(e), response.status_code)            
+            if response != None:
+                print("Reason: " + response.text, file=sys.stderr)
+            sys.exit(1)
+
+    def handle_job_error_status(self, response_json):
+        self.printHttpErr("GET upload status failed. " + response_json["statusDetails"]["message"], "200")
+        sys.exit(1)
+
+    def handle_job_rejected_status(self, response_json):
+        msgLines = ["Export failed.  Dataset had validation problems:"]
+        for general in response_json["statusDetails"]["general"]:
+            msgLines.append(general)
+        for key in response_json["statusDetails"]["byKey"]:
+            msgLines.append(key + ": " + response_json["statusDetails"]["byKey"][key])
+        print('\n'.join(msgLines), file=sys.stderr)
+        sys.exit(1)
+
+    def printHttpErr(self, msg, status_code):
+        print("Http Error (" + status_code + "): " + msg, file=sys.stderr)            
+
     def output_success(self):
         header = "<html><body><h1>Good news!</h1><br />"
         msg = """
-        <h2>Results of the VEuPathDB Export Tool<br />Bigwig Files to VEuPathDB</h2>
-        <h3>Your set of bigwig files was exported from Galaxy to your account in VEuPathDB.
-         For file access and to view in GBrowse, go to My Data Sets in the appropriate VEuPathDB site:
-        </h3><br />
-        Go to the appropriate VEuPathDB site (links below) to see it (and all your User Datasets):<br \>
+        <h2>Your export to VEuPathDB is complete<h2>.
+         View the exported dataset by visiting the My Data Sets page in the appropriate VEuPathDB site
+        <br />
         <a href='http://amoebadb.org/amoeba/app/workspace/datasets'>AmoebaDB</a><br />
         <a href='http://cryptodb.org/cryptodb/app/workspace/datasets'>CryptoDB</a><br />
         <a href='http://fungidb.org/fungidb/app/workspace/datasets'>FungiDB</a><br />
@@ -401,14 +243,38 @@ class Export:
         <a href='http://tritrypdb.org/tritrypdb/app/workspace/datasets'>TriTrypDB</a><br />
         </body></html>
         """
-        with open(self._output, 'w') as file:
+        with open(self._stdArgsBundle.output, 'w') as file:
             file.write("%s%s" % (header,msg))
 
+    def identify_projects(self):
+        """
+        An abstract method to be addressed by a specialized export tool that furnishes a VEuPathDB project list.
+        :return: The project list to be returned should look as follows:
+        [project1, project2, ... ]
+        At least one valid VEuPathDB project must be listed
+        """
+        raise NotImplementedError(
+            "The method 'identify_project(self)' needs to be implemented in the specialized export module.")
+
+    def identify_dataset_files(self):
+        """
+        An abstract method to be addressed by a specialized export tool that furnishes a json list
+        containing the dataset data files and the VEuPathDB file names they must have in the tarball.
+        :return: The dataset file list to be returned should look as follows:
+        [dataset file1, dataset file2, ... ]
+        where each dataset file is written as a json object as follows:
+        {
+          "name":<filename that VEuPathDB expects>,
+          "path":<Galaxy path to the dataset file>
+        At least one valid VEuPathDB dataset file must be listed
+        """
+        raise NotImplementedError(
+            "The method 'identify_dataset_file(self)' needs to be implemented in the specialized export module.")
 
 
-class ValidationException(Exception):
+class SystemException(Exception):
     """
-    This represents the exception reported when a call to a validation script returns a data error.
+    An unexpected system error.
     """
     pass
 
@@ -418,3 +284,15 @@ class TransferException(Exception):
     This represents the exception reported when the export of a dataset to the iRODS system returns a failure.
     """
     pass
+
+# expect email of the form: sfischer.67546@veupathdb.org (where the number is the user ID)
+def getWdkUserId(rawUserEmail):
+    user_email = rawUserEmail.strip()
+    # WDK user id is derived from the user email
+    if not re.match(r'.+\.\d+@veupathdb.org$', user_email, flags=0):
+        raise SystemException("The user email " + str(user_email) + " is not valid for the use of this tool.")
+    galaxy_user = user_email.split("@")[0]
+    return galaxy_user[galaxy_user.rfind(".") + 1:]
+
+        
+
